add a wrapper around the containerd.New call to fix and pass the proper npipe connector
